+++ TSOClient/FSO.Server.Database/DA/Objects/IObjects.cs
@@ -29,5 +29,6 @@ namespace FSO.Server.Database.DA.Objects
         int GetSpecificObjectAttribute(uint objectID, int index);
         void SetObjectAttributes(List<DbObjectAttribute> attrs);
         int TotalObjectAttributes(uint guid, int index);
+        DbObjectNetWorth GetNetWorth(uint owner_id);
     }
 }